Move GenesisConfig from linera-client to linera-core
Move GenesisConfig and its Error type to linera-core::genesis_config.
Re-export from linera-client::config for backward compatibility.
The old GenesisConfig::new(CommitteeConfig, ...) is replaced by
CommitteeConfig::into_genesis(...) on the client side, while the
core GenesisConfig::new takes a Committee directly.
