+++ linera-client/src/unit_tests/chain_listener.rs
@@ -98,7 +98,7 @@ async fn test_chain_listener() -> anyhow::Result<()> {
     let chain_id0 = client0.chain_id();
     let client1 = builder.add_root_chain(1, Amount::ONE).await?;
     // Start a chain listener for chain 0 with a new key.
-    let genesis_config = GenesisConfig::new_testing(&builder);
+    let genesis_config = GenesisConfig::new_for_testing(&builder);
     let admin_chain_id = genesis_config.admin_chain_id();
     let storage = builder.make_storage().await?;
     let epoch0 = client0.chain_info().await?.epoch;
@@ -208,7 +208,7 @@ async fn test_chain_listener_follow_only() -> anyhow::Result<()> {
     let chain_a_id = chain_a.chain_id();
     let chain_b_id = chain_b.chain_id();
 
-    let genesis_config = GenesisConfig::new_testing(&builder);
+    let genesis_config = GenesisConfig::new_for_testing(&builder);
     let admin_chain_id = genesis_config.admin_chain_id();
     let storage = builder.make_storage().await?;
     let chain_a_info = chain_a.chain_info().await?;
@@ -368,7 +368,7 @@ async fn test_chain_listener_admin_chain() -> anyhow::Result<()> {
     let storage_builder = MemoryStorageBuilder::default();
     let mut builder = TestBuilder::new(storage_builder, 4, 1, signer.clone()).await?;
     let client0 = builder.add_root_chain(0, Amount::ONE).await?;
-    let genesis_config = GenesisConfig::new_testing(&builder);
+    let genesis_config = GenesisConfig::new_for_testing(&builder);
     let admin_chain_id = genesis_config.admin_chain_id();
     let storage = builder.make_storage().await?;
 
@@ -443,7 +443,7 @@ async fn test_chain_listener_listen_command_adds_chains_to_wallet() -> anyhow::R
     let client0 = builder.add_root_chain(0, Amount::ONE).await?;
     let chain_id0 = client0.chain_id();
 
-    let genesis_config = GenesisConfig::new_testing(&builder);
+    let genesis_config = GenesisConfig::new_for_testing(&builder);
     let admin_chain_id = genesis_config.admin_chain_id();
     let storage = builder.make_storage().await?;
 
@@ -560,7 +560,7 @@ async fn test_listener_uses_autosigner_for_incoming_messages() -> anyhow::Result
         ))
         .await?;
 
-    let genesis_config = GenesisConfig::new_testing(&builder);
+    let genesis_config = GenesisConfig::new_for_testing(&builder);
     let admin_chain_id = genesis_config.admin_chain_id();
     let storage = builder.make_storage().await?;
